% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R
\name{acceptAlert}
\alias{acceptAlert}
\title{acceptAlert}
\usage{
acceptAlert(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
acceptAlert
}
\seealso{
Other userPrompts functions: \code{\link{dismissAlert}},
  \code{\link{getAlertText}}, \code{\link{sendAlertText}}
}

