% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{acceptAlertOld}
\alias{acceptAlertOld}
\title{acceptAlertOld}
\usage{
acceptAlertOld(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{acceptAlertOld}
}
\examples{

}
\seealso{
Other oldMethods functions: \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{sendAlertTextOld}}
}

