% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteDriver.R
\name{addCookie}
\alias{addCookie}
\title{addCookie}
\usage{
addCookie(remDr, name, value, path = NULL, domain = NULL, secure = FALSE,
  httpOnly = NULL, expiry = NULL, ...)
}
\arguments{
\item{remDr}{}
}
\description{
addCookie
}

