% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{back}
\alias{back}
\title{back}
\usage{
back(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{back}
}
\examples{

}
\seealso{
Other navigation functions: \code{\link{forward}},
  \code{\link{getCurrentUrl}}, \code{\link{getTitle}},
  \code{\link{go}}, \code{\link{refresh}}
}

