% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R, R/commandContextsDoc.R
\name{closeWindow}
\alias{closeWindow}
\title{closeWindow}
\usage{
closeWindow(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
closeWindow
}
\seealso{
Other commandContexts functions: \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

