% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R, R/cookiesDoc.R
\name{deleteCookie}
\alias{deleteCookie}
\title{deleteCookie}
\usage{
deleteCookie(remDr, name = NULL, ...)
}
\arguments{
\item{remDr}{remote driver doc}

\item{name}{}
}
\description{
deleteCookie
}
\seealso{
Other cookies functions: \code{\link{addCookie}},
  \code{\link{deleteAllCookies}},
  \code{\link{getAllCookies}}, \code{\link{getNamedCookie}}
}

