% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{deleteSession}
\alias{deleteSession}
\title{deleteSession}
\usage{
deleteSession(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
ret9 : .e$sessionId[[remDr$drvID]] <- NULL;invisible(remDr)
}
\description{
\code{deleteSession}
}
\examples{

}
\seealso{
Other sessions functions: \code{\link{newSession}},
  \code{\link{setTimeout}}
}

