% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R, R/userPromptsDoc.R
\name{dismissAlert}
\alias{dismissAlert}
\title{dismissAlert}
\usage{
dismissAlert(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
dismissAlert
}
\seealso{
Other userPrompts functions: \code{\link{acceptAlert}},
  \code{\link{getAlertText}}, \code{\link{sendAlertText}}
}

