% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R, R/elementInteractionDoc.R
\name{elementClear}
\alias{elementClear}
\title{elementClear}
\usage{
elementClear(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
ret7 : invisible(webElem)
}
\description{
\code{elementClear}
}
\examples{

}
\seealso{
Other elementInteraction functions: \code{\link{elementClick}},
  \code{\link{elementSendKeys}}
}

