% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R, R/elementInteractionDoc.R
\name{elementClick}
\alias{elementClick}
\title{elementClick}
\usage{
elementClick(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
ret7 : invisible(webElem)
}
\description{
\code{elementClick}
}
\examples{

}
\seealso{
Other elementInteraction functions: \code{\link{elementClear}},
  \code{\link{elementSendKeys}}
}

