% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R, R/elementInteractionDoc.R
\name{elementSendKeys}
\alias{elementSendKeys}
\title{elementSendKeys}
\usage{
elementSendKeys(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
elementSendKeys
}
\seealso{
Other elementInteraction functions: \code{\link{elementClear}},
  \code{\link{elementClick}}
}

