% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{errorContent}
\alias{errorContent}
\title{Returns the content from remote webdriver}
\usage{
errorContent()
}
\value{
returns content see \code{\link{content}}
}
\description{
\code{errorContent} returns the content from the remote webdriver on an error.
}

