% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{errorResponse}
\alias{errorResponse}
\title{Return the response from remote webdriver}
\usage{
errorResponse()
}
\value{
returns response see \code{\link{VERB}}. Headers, request etc. can be examined from this object.
}
\description{
\code{errorResponse} returns the response from the remote webdriver on an error.
}
\examples{
\dontrun{
remDr <- remoteDr()
remDr \%>\% findElement("name", "sdofnsdofk")
errorResponse()
}
}

