% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R, R/documentHandlingDoc.R
\name{executeAsyncScript}
\alias{executeAsyncScript}
\title{executeAsyncScript}
\usage{
executeAsyncScript(remDr, script, args = list(), replace = TRUE, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{replace}{}
}
\value{
ret6 : if(replace){testWebElement(res$value, remDr)}else{res$value}
}
\description{
\code{executeAsyncScript}
}
\examples{

}
\seealso{
Other documentHandling functions: \code{\link{executeScript}},
  \code{\link{getPageSource}}
}

