% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R, R/documentHandlingDoc.R
\name{executeScript}
\alias{executeScript}
\title{Execute JavaScript on browser.}
\usage{
executeScript(remDr, script, args = list(), replace = TRUE, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{script}{character: The script to inject.}

\item{args}{The arguments of the script as a list.}

\item{replace}{logical: If TRUE any elements identify as web Elements are converted to such.}
}
\value{
If replace is TRUE parses the response from the server for web Elements and converts as such.
   Otherwise the content returned is assummed to be a simple list.
}
\description{
\code{executeScript}  Inject a snippet of JavaScript into the page for execution in the context
   of the currently selected frame. The executed script is assumed to be
   synchronous and the result of evaluating the script will be returned.
}
\examples{

}
\seealso{
Other documentHandling functions: \code{\link{executeAsyncScript}},
  \code{\link{getPageSource}}
}

