% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{executeScriptOld}
\alias{executeScriptOld}
\title{executeScriptOld}
\usage{
executeScriptOld(remDr, script, args = list(), replace = TRUE, ...)
}
\arguments{
\item{remDr}{remote driver doc}

\item{replace}{}
}
\description{
executeScriptOld
}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{sendAlertTextOld}}
}

