% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteDriver.R
\name{findElements}
\alias{findElements}
\title{findElements}
\usage{
findElements(remDr, using = c("xpath", "css selector", "id", "name",
  "tag name", "class name", "link text", "partial link text"), value, ...)
}
\arguments{
\item{remDr}{}
}
\description{
findElements
}

