% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementRetrieval.R, R/elementRetrievalDoc.R
\name{findElementsFromElement}
\alias{findElementsFromElement}
\title{findElementsFromElement}
\usage{
findElementsFromElement(webElem, using = c("xpath", "css selector", "id",
  "name", "tag name", "class name", "link text", "partial link text"), value,
  ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{value}{}
}
\value{
ret11 : invisible(lapply(res$value, wbElement, remDr = webElem$remDr))
}
\description{
\code{findElementsFromElement}
}
\examples{

}
\seealso{
Other elementRetrieval functions: \code{\link{findElementFromElement}},
  \code{\link{findElements}}, \code{\link{findElement}},
  \code{\link{getActiveElement}}
}

