% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R
\name{fullscreenWindow}
\alias{fullscreenWindow}
\title{fullscreenWindow}
\usage{
fullscreenWindow(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
fullscreenWindow
}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

