% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R, R/userPromptsDoc.R
\name{getAlertText}
\alias{getAlertText}
\title{getAlertText}
\usage{
getAlertText(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
ret2 : res$value NEED TO FILL IN specifics for each function
}
\description{
\code{getAlertText}
}
\examples{

}
\seealso{
Other userPrompts functions: \code{\link{acceptAlert}},
  \code{\link{dismissAlert}}, \code{\link{sendAlertText}}
}

