% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{getCookie}
\alias{getCookie}
\title{getCookie}
\usage{
getCookie(remDr, name = NULL, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getCookie
}
\seealso{
Other cookies functions: \code{\link{addCookie}},
  \code{\link{deleteAllCookies}},
  \code{\link{deleteCookie}}
}

