% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{getCurrentUrl}
\alias{getCurrentUrl}
\title{Retrieve the URL of the current page.}
\usage{
getCurrentUrl(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
A character string is returned giving the current page URL.
}
\description{
\code{getCurrentUrl} Retrieve the URL of the current page.
}
\examples{

}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getTitle}},
  \code{\link{go}}, \code{\link{refresh}}
}

