% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementCssValue}
\alias{getElementCssValue}
\title{getElementCssValue}
\usage{
getElementCssValue(webElem, propertyName, ...)
}
\arguments{
\item{webElem}{}
}
\description{
getElementCssValue
}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

