% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementProperty}
\alias{getElementProperty}
\title{Query the value of an elements property.}
\usage{
getElementProperty(webElem, property, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{property}{The property to query as a character string}
}
\value{
The value of the elements specified property.
}
\description{
\code{getElementProperty} Query the value of an elements property.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

