% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementRect}
\alias{getElementRect}
\title{Return the dimensions and coordinates of an element}
\usage{
getElementRect(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
The returned value is a list including the following members:
\describe{
  \item{x}{X axis position of the top-left corner of the web element relative to the current browsing context’s document element in CSS reference pixels.}
  \item{y}{Y axis position of the top-left corner of the web element relative to the current browsing context’s document element in CSS reference pixels.}
  \item{height}{Height of the web element’s bounding rectangle in CSS reference pixels. }
  \item{width}{Width of the web element’s bounding rectangle in CSS reference pixels. }
}
}
\description{
\code{getElementRect} The getElementRect fuinction returns the dimensions and coordinates of the given web element.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

