% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementTagName}
\alias{getElementTagName}
\title{Query for an element's tag name.}
\usage{
getElementTagName(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
The element's tag name, as a lowercase character string.
}
\description{
\code{getElementTagName} Query for an element's tag name.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

