% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R, R/documentHandlingDoc.R
\name{getPageSource}
\alias{getPageSource}
\title{getPageSource}
\usage{
getPageSource(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
ret3 : read_html(res$value)
}
\description{
\code{getPageSource}
}
\examples{

}
\seealso{
Other documentHandling functions: \code{\link{executeAsyncScript}},
  \code{\link{executeScript}}
}

