% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{getTitle}
\alias{getTitle}
\title{getTitle}
\usage{
getTitle(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
ret2 : res$value NEED TO FILL IN specifics for each function
}
\description{
\code{getTitle}
}
\examples{

}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getCurrentUrl}},
  \code{\link{go}}, \code{\link{refresh}}
}

