% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{getWindowHandlesOld}
\alias{getWindowHandlesOld}
\title{getWindowHandlesOld}
\usage{
getWindowHandlesOld(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getWindowHandlesOld
}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{sendAlertTextOld}}
}

