% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{getWindowSizeOld}
\alias{getWindowSizeOld}
\title{Get window size}
\usage{
getWindowSizeOld(remDr, handle = "current", ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{handle}{Handle of the window to query. If handle = "current" (the default) the current window is used.}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
The width and height of the window are returned as elements in a list.
}
\description{
\code{getWindowSizeOld} Get the size of the specified window. Uses the old JSONwireprotocol end points.
}
\examples{
\dontrun{
# functions in this group are using the old JSONwireprotocol end points
}

}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{getWindowPositionOld}},
  \code{\link{maximizeWindowOld}},
  \code{\link{sendAlertTextOld}},
  \code{\link{setWindowPositionOld}},
  \code{\link{setWindowSizeOld}}
}

