% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{go}
\alias{go}
\title{go}
\usage{
go(remDr, url, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{url}{}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{go}
}
\examples{

}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getCurrentUrl}},
  \code{\link{getTitle}}, \code{\link{refresh}}
}

