% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{isElementEnabled}
\alias{isElementEnabled}
\title{Determine if an element is currently enabled.}
\usage{
isElementEnabled(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
A logical value is returned indicating whether the elemnent is enabled.
}
\description{
\code{isElementEnabled} Determine if an element is currently enabled.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementSelected}}
}

