% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{isElementSelected}
\alias{isElementSelected}
\title{Determine if an element is currently selected.}
\usage{
isElementSelected(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
A logical value is returned indicating whether the elemnent is selected.
}
\description{
\code{isElementSelected} Determines if an OPTION element, or an INPUT element of type checkbox or radiobutton is currently selected.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}}
}

