% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R
\name{maximizeWindow}
\alias{maximizeWindow}
\title{maximizeWindow}
\usage{
maximizeWindow(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
maximizeWindow
}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

