% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{maximizeWindowOld}
\alias{maximizeWindowOld}
\title{Maximize the current window.}
\usage{
maximizeWindowOld(remDr, handle = "current", ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{handle}{Handle of the window to query. If handle = "current"
(the default) the current window is used.}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{maximizeWindowOld} Maximize the specified window if not already
   maximized.
}
\examples{
\dontrun{
# functions in this group are using the old JSONwireprotocol end points
}

}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{getWindowPositionOld}},
  \code{\link{getWindowSizeOld}},
  \code{\link{sendAlertTextOld}},
  \code{\link{setWindowPositionOld}},
  \code{\link{setWindowSizeOld}}
}

