% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{newSession}
\alias{newSession}
\title{newSession}
\usage{
newSession(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
ret8 : remDr$sessionInfo <- res$value;invisible(remDr)
}
\description{
\code{newSession}
}
\examples{

}
\seealso{
Other sessions functions: \code{\link{deleteSession}},
  \code{\link{setTimeout}}
}

