% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R, R/interactionsDoc.R
\name{performActions}
\alias{performActions}
\title{performActions}
\usage{
performActions(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{performActions}
}
\examples{

}
\seealso{
Other interactions functions: \code{\link{releasingActions}}
}

