% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{queryDriver}
\alias{queryDriver}
\title{Send a query to remote Driver.}
\usage{
queryDriver(verb = GET, url, source, drvID, ...)
}
\arguments{
\item{verb}{The http method to use. See \code{\link{VERB}}}

\item{url}{The url of the remote server endpoint.}

\item{source}{The name of the seleniumPipes function that called queryDriver.}

\item{drvID}{The driver id of the session as given by an object of class "remoteDr"}

\item{...}{additonal arguments}
}
\value{
The contents of the response from the remote server. See \code{\link{content}} for details.
}
\description{
\code{queryDriver} A function to send a query to a remote driver. Intended for seleniumPipes
   internal use mainly.
}
\examples{

}

