% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R, R/interactionsDoc.R
\name{releasingActions}
\alias{releasingActions}
\title{releasingActions}
\usage{
releasingActions(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
releasingActions
}
\seealso{
Other interactions functions: \code{\link{performActions}}
}

