% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{seleniumPipes}
\alias{seleniumPipes}
\alias{seleniumPipes-package}
\title{Implements the W3C webdriver specification.}
\description{
Implements the W3C webdriver specification.
}
\section{Package options}{


seleniumPipes uses the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{seleniumPipes_display_screenshot}: Logical value indicating whether to display
     PNG returned by \code{\link{takeScreenshot}} and \code{\link{takeElementScreenshot}}.
     Defaults to TRUE

  \item \code{seleniumPipes_retry}: An integer giving the number of time to try calling an endpoint
     on the Selenium Server. Defaults to 3 attempts

  \item \code{seleniumPipes_retry_delay}: An integer detailing the delay between attempts to
     call a failing endpoint on the Selenium Server. Defaults to 5000 miliseconds = 5 seconds.

  \item \code{seleniumPipes_SL}: A logical value which acts as a flag indicating whether SauiceLabs is being
     used for package testing.

  \item \code{seleniumPipes_selOptions}: A list used to store options to pass to \code{\link{remoteDr}}
     when running tests.

  \item \code{seleniumPipes_sauceID}: A character used to store remote session ids when running SauceLab tests on the
     package.

}
}

