% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R, R/userPromptsDoc.R
\name{sendAlertText}
\alias{sendAlertText}
\title{sendAlertText}
\usage{
sendAlertText(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
sendAlertText
}
\seealso{
Other userPrompts functions: \code{\link{acceptAlert}},
  \code{\link{dismissAlert}}, \code{\link{getAlertText}}
}

