% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{setTimeout}
\alias{setTimeout}
\title{setTimeout}
\usage{
setTimeout(remDr, type = "page load", milliseconds = 10000, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{milliseconds}{}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{setTimeout}
}
\examples{

}
\seealso{
Other sessions functions: \code{\link{deleteSession}},
  \code{\link{newSession}}
}

