% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{status}
\alias{status}
\title{Get remote end status.}
\usage{
status(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
ret2 : res$value NEED TO FILL IN specifics for each function
}
\description{
\code{status} The Status command returns information about whether a
   remote end is in a state in which it can create a new session.
   This is represented by the ready property of the response body,
   which has a value of false if attempting to create a session at
   the current time would fail.
}
\examples{
\dontrun{
  # start a driver without opening a browser
  remDr <- remoteDr(newSession = FALSE)

  # open a broswer
  remDr \%>\% newSession

  # set timeout on waiting for elements
  remDr \%>\% setTimeout(type = "implicit", 5000)

  # close Session
  remDr \%>\% deleteSession

}

}
\seealso{
Other sessions functions: \code{\link{deleteSession}},
  \code{\link{getTimeouts}}, \code{\link{newSession}},
  \code{\link{setTimeouts}}
}

