% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenCapture.R, R/screenCaptureDoc.R
\name{takeScreenshot}
\alias{takeScreenshot}
\title{takeScreenshot}
\usage{
takeScreenshot(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
takeScreenshot
}
\seealso{
Other screenCapture functions: \code{\link{takeElementScreenshot}}
}

