% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R, R/userPromptsDoc.R
\name{acceptAlert}
\alias{acceptAlert}
\title{acceptAlert}
\usage{
acceptAlert(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{acceptAlert}
}
\examples{

}
\seealso{
Other userPrompts functions: \code{\link{dismissAlert}},
  \code{\link{getAlertText}}, \code{\link{sendAlertText}}
}

