% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R
\name{acceptAlertOld}
\alias{acceptAlertOld}
\title{acceptAlertOld}
\usage{
acceptAlertOld(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
acceptAlertOld
}
\seealso{
Other oldMethods functions: \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{sendAlertTextOld}}
}

