% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{acceptAlertOld}
\alias{acceptAlertOld}
\title{Accept alert}
\usage{
acceptAlertOld(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{acceptAlertOld} accept a JavaScript alert This uses the old
   JSONwireprotocol endpoints.
}
\examples{
\dontrun{
# functions in this group are using the old JSONwireprotocol end points
}

}
\seealso{
Other oldMethods functions: \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{getWindowPositionOld}},
  \code{\link{getWindowSizeOld}},
  \code{\link{maximizeWindowOld}},
  \code{\link{sendAlertTextOld}},
  \code{\link{setWindowPositionOld}},
  \code{\link{setWindowSizeOld}}
}

