% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R, R/cookiesDoc.R
\name{addCookie}
\alias{addCookie}
\title{addCookie}
\usage{
addCookie(remDr, name, value, path = NULL, domain = NULL, secure = FALSE,
  httpOnly = NULL, expiry = NULL, ...)
}
\arguments{
\item{remDr}{remote driver doc}

\item{expiry}{}
}
\description{
addCookie
}
\seealso{
Other cookies functions: \code{\link{deleteAllCookies}},
  \code{\link{deleteCookie}}, \code{\link{getAllCookies}},
  \code{\link{getNamedCookie}}
}

