% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{back}
\alias{back}
\title{back}
\usage{
back(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
back
}
\seealso{
Other navigation functions: \code{\link{forward}},
  \code{\link{getCurrentUrl}}, \code{\link{getTitle}},
  \code{\link{go}}, \code{\link{refresh}}
}

