% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{deleteAllCookies}
\alias{deleteAllCookies}
\title{deleteAllCookies}
\usage{
deleteAllCookies(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
deleteAllCookies
}
\seealso{
Other cookies functions: \code{\link{addCookie}},
  \code{\link{deleteCookie}}, \code{\link{getAllCookies}},
  \code{\link{getNamedCookie}}
}

