% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{deleteSession}
\alias{deleteSession}
\title{deleteSession}
\usage{
deleteSession(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
deleteSession
}
\seealso{
Other sessions functions: \code{\link{newSession}},
  \code{\link{setTimeout}}
}

