% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{dismissAlertOld}
\alias{dismissAlertOld}
\title{dismissAlertOld}
\usage{
dismissAlertOld(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
dismissAlertOld
}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{sendAlertTextOld}}
}

