% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R, R/elementInteractionDoc.R
\name{elementClear}
\alias{elementClear}
\title{Clear an elements text value.}
\usage{
elementClear(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
invisible(webElem): An object of class "wElement" is invisibly returned.
   A webElement object see \code{\link{wbElement}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{elementClear} Clear a TEXTAREA or text INPUT element's value.
}
\examples{

}
\seealso{
Other elementInteraction functions: \code{\link{elementClick}},
  \code{\link{elementSendKeys}}
}

