% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R
\name{elementClear}
\alias{elementClear}
\title{elementClear}
\usage{
elementClear(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
elementClear
}
\seealso{
Other elementInteraction functions: \code{\link{elementClick}},
  \code{\link{elementSendKeys}}
}

