% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R
\name{elementClick}
\alias{elementClick}
\title{elementClick}
\usage{
elementClick(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
elementClick
}
\seealso{
Other elementInteraction functions: \code{\link{elementClear}},
  \code{\link{elementSendKeys}}
}

