% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementInteraction.R, R/elementInteractionDoc.R
\name{elementSendKeys}
\alias{elementSendKeys}
\title{elementSendKeys}
\usage{
elementSendKeys(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
ret7 : invisible(webElem)
}
\description{
\code{elementSendKeys}
}
\examples{

}
\seealso{
Other elementInteraction functions: \code{\link{elementClear}},
  \code{\link{elementClick}}
}

