% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteDriver.R
\name{executeAsyncScript}
\alias{executeAsyncScript}
\title{executeAsyncScript}
\usage{
executeAsyncScript(remDr, script, args = list(), replace = TRUE, ...)
}
\arguments{
\item{remDr}{}
}
\description{
executeAsyncScript
}

