% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R, R/documentHandlingDoc.R
\name{executeAsyncScript}
\alias{executeAsyncScript}
\title{Execute JavaScript asynchronously on browser.}
\usage{
executeAsyncScript(remDr, script, args = list(), replace = TRUE, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{script}{character: The script to inject.}

\item{args}{The arguments of the script as a list.}

\item{replace}{logical: If TRUE any elements identify as web Elements are converted to such.}
}
\value{
If replace is TRUE parses the response from the server for web Elements and converts as such.
   Otherwise the content returned is assummed to be a simple list.
}
\description{
\code{executeAsyncScript} Inject a snippet of JavaScript into the page for asynchronous execution in
   the context of the currently selected frame.
}
\details{
The driver will pass a callback as the last argument to the snippet, and
   block until the callback is invoked.
}
\examples{

}
\seealso{
Other documentHandling functions: \code{\link{executeScript}},
  \code{\link{getPageSource}}
}

