% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R
\name{executeScript}
\alias{executeScript}
\title{executeScript}
\usage{
executeScript(remDr, script, args = list(), replace = TRUE, ...)
}
\arguments{
\item{remDr}{}
}
\description{
executeScript
}
\seealso{
Other documentHandling functions: \code{\link{executeAsyncScript}},
  \code{\link{getPageSource}}
}

