% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementRetrieval.R
\name{findElement}
\alias{findElement}
\title{findElement}
\usage{
findElement(remDr, using = c("xpath", "css selector", "id", "name",
  "tag name", "class name", "link text", "partial link text"), value, ...)
}
\arguments{
\item{remDr}{}
}
\description{
findElement
}
\seealso{
Other elementRetrieval functions: \code{\link{findElementFromElement}},
  \code{\link{findElementsFromElement}},
  \code{\link{findElements}},
  \code{\link{getActiveElement}}
}

