% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementRetrieval.R, R/elementRetrievalDoc.R
\name{findElementFromElement}
\alias{findElementFromElement}
\title{findElementFromElement}
\usage{
findElementFromElement(webElem, using = c("xpath", "css selector", "id",
  "name", "tag name", "class name", "link text", "partial link text"), value,
  ...)
}
\arguments{
\item{webElem}{}
}
\description{
findElementFromElement
}
\seealso{
Other elementRetrieval functions: \code{\link{findElementsFromElement}},
  \code{\link{findElements}}, \code{\link{findElement}},
  \code{\link{getActiveElement}}
}

