% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementRetrieval.R, R/elementRetrievalDoc.R
\name{findElements}
\alias{findElements}
\title{findElements}
\usage{
findElements(remDr, using = c("xpath", "css selector", "id", "name",
  "tag name", "class name", "link text", "partial link text"), value, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{value}{}
}
\value{
ret5 : invisible(lapply(res$value, wbElement, remDr = remDr))
}
\description{
\code{findElements}
}
\examples{

}
\seealso{
Other elementRetrieval functions: \code{\link{findElementFromElement}},
  \code{\link{findElementsFromElement}},
  \code{\link{findElement}}, \code{\link{getActiveElement}}
}

