% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{forward}
\alias{forward}
\title{Navigate forwards}
\usage{
forward(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{forward} Navigate forwards in the browser history, if possible.
}
\examples{

}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{getCurrentUrl}}, \code{\link{getTitle}},
  \code{\link{go}}, \code{\link{refresh}}
}

