% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementRetrieval.R, R/elementRetrievalDoc.R
\name{getActiveElement}
\alias{getActiveElement}
\title{getActiveElement}
\usage{
getActiveElement(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
ret4 : invisible(wbElement(res$value, remDr))
}
\description{
\code{getActiveElement}
}
\examples{

}
\seealso{
Other elementRetrieval functions: \code{\link{findElementFromElement}},
  \code{\link{findElementsFromElement}},
  \code{\link{findElements}}, \code{\link{findElement}}
}

