% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementRetrieval.R, R/elementRetrievalDoc.R
\name{getActiveElement}
\alias{getActiveElement}
\title{getActiveElement}
\usage{
getActiveElement(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getActiveElement
}
\seealso{
Other elementRetrieval functions: \code{\link{findElementFromElement}},
  \code{\link{findElementsFromElement}},
  \code{\link{findElements}}, \code{\link{findElement}}
}

