% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R, R/userPromptsDoc.R
\name{getAlertText}
\alias{getAlertText}
\title{getAlertText}
\usage{
getAlertText(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
getAlertText
}
\seealso{
Other userPrompts functions: \code{\link{acceptAlert}},
  \code{\link{dismissAlert}}, \code{\link{sendAlertText}}
}

