% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userPrompts.R
\name{getAlertText}
\alias{getAlertText}
\title{getAlertText}
\usage{
getAlertText(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getAlertText
}
\seealso{
Other userPrompts functions: \code{\link{acceptAlert}},
  \code{\link{dismissAlert}}, \code{\link{sendAlertText}}
}

