% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R
\name{getAlertTextOld}
\alias{getAlertTextOld}
\title{getAlertTextOld}
\usage{
getAlertTextOld(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getAlertTextOld
}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{sendAlertTextOld}}
}

