% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{getAlertTextOld}
\alias{getAlertTextOld}
\title{Get alert text}
\usage{
getAlertTextOld(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
The alert text is returned as a string.
}
\description{
\code{getAlertTextOld} Get the text from a JavaScript alert. This uses the old JSONwireprotocol endpoints.
}
\examples{
\dontrun{
# functions in this group are using the old JSONwireprotocol end points
}

}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{getWindowPositionOld}},
  \code{\link{getWindowSizeOld}},
  \code{\link{maximizeWindowOld}},
  \code{\link{sendAlertTextOld}},
  \code{\link{setWindowPositionOld}},
  \code{\link{setWindowSizeOld}}
}

