% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R, R/cookiesDoc.R
\name{getAllCookies}
\alias{getAllCookies}
\title{getAllCookies}
\usage{
getAllCookies(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getAllCookies
}
\seealso{
Other cookies functions: \code{\link{addCookie}},
  \code{\link{deleteAllCookies}},
  \code{\link{deleteCookie}}, \code{\link{getNamedCookie}}
}

