% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{getCurrentUrl}
\alias{getCurrentUrl}
\title{getCurrentUrl}
\usage{
getCurrentUrl(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getCurrentUrl
}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getTitle}},
  \code{\link{go}}, \code{\link{refresh}}
}

