% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementAttribute}
\alias{getElementAttribute}
\title{Get the value of an element's attribute.}
\usage{
getElementAttribute(webElem, attribute, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{attribute}{The attribute to query as a character string.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
The value of the attribute, or null if it is not set on the element.
}
\description{
\code{getElementAttribute} Get the value of an element's attribute.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

