% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementAttribute}
\alias{getElementAttribute}
\title{getElementAttribute}
\usage{
getElementAttribute(webElem, attribute, ...)
}
\arguments{
\item{webElem}{}
}
\description{
getElementAttribute
}
\seealso{
Other elementState functions: \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

