% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementCssValue}
\alias{getElementCssValue}
\title{Query the value of an element's computed CSS property.}
\usage{
getElementCssValue(webElem, propertyName, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{propertyName}{The property to query as a character string}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
The value of the specified CSS property.
}
\description{
\code{getElementCssValue} Query the value of an element's computed CSS property.
 The CSS property to query should be specified using the CSS property name,
 not the JavaScript property name (e.g. background-color instead of backgroundColor).
}
\examples{
\dontrun{
  remDr <- remoteDr()
  remDr \%>\% go("http://www.google.com/ncr")

  # select the search box
  searchElem <- remDr \%>\% findElement("name", "q")

  # get the name of our element
  searchElem \%>\% getElementAttribute("name")

  # get the css background-color
  searchElem \%>\% getElementCssValue("background-color")
  blueScript <- "arguments[0].style.backgroundColor = 'blue';"
  remDr \%>\% executeScript(blueScript, list(searchElem))
  searchElem \%>\% getElementCssValue("background-color")

  # get the javascript property
  # searchElem \%>\% getElementProperty("backgroundColor")

  # get dimensions
  searchElem \%>\% getElementRect

  searchElem \%>\% getElementTagName

  # get some text and return it
  remDr \%>\% go("http://r-project.org")
  remDr \%>\% findElement("css", "h1") \%>\% getElementText

  # close browser
  remDr \%>\% deleteSession
}

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

