% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementCssValue}
\alias{getElementCssValue}
\title{Query the value of an element's computed CSS property.}
\usage{
getElementCssValue(webElem, propertyName, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{propertyName}{The property to query as a character string}
}
\value{
The value of the specified CSS property.
}
\description{
\code{getElementCssValue} Query the value of an element's computed CSS property.
 The CSS property to query should be specified using the CSS property name,
 not the JavaScript property name (e.g. background-color instead of backgroundColor).
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

