% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R
\name{getElementProperty}
\alias{getElementProperty}
\title{getElementProperty}
\usage{
getElementProperty(webElem, property, ...)
}
\arguments{
\item{webElem}{}
}
\description{
getElementProperty
}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

