% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementProperty}
\alias{getElementProperty}
\title{getElementProperty}
\usage{
getElementProperty(webElem, property, ...)
}
\arguments{
\item{webElem}{web Element doc}

\item{property}{}
}
\value{
ret2 : res$value NEED TO FILL IN specifics for each function
}
\description{
\code{getElementProperty}
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

