% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R
\name{getElementRect}
\alias{getElementRect}
\title{getElementRect}
\usage{
getElementRect(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
getElementRect
}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

