% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementTagName}
\alias{getElementTagName}
\title{getElementTagName}
\usage{
getElementTagName(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
ret2 : res$value NEED TO FILL IN specifics for each function
}
\description{
\code{getElementTagName}
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

