% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementTagName}
\alias{getElementTagName}
\title{getElementTagName}
\usage{
getElementTagName(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
getElementTagName
}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementText}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

