% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R
\name{getElementText}
\alias{getElementText}
\title{getElementText}
\usage{
getElementText(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
getElementText
}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

