% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R, R/elementStateDoc.R
\name{getElementText}
\alias{getElementText}
\title{Returns the visible text for the element.}
\usage{
getElementText(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
The visible text for an element is returned as a character string.
}
\description{
\code{getElementText} Returns the visible text for the element.
}
\examples{

}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{isElementEnabled}},
  \code{\link{isElementSelected}}
}

