% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{getNamedCookie}
\alias{getNamedCookie}
\title{getNamedCookie}
\usage{
getNamedCookie(remDr, name = NULL, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getNamedCookie
}
\seealso{
Other cookies functions: \code{\link{addCookie}},
  \code{\link{deleteAllCookies}},
  \code{\link{deleteCookie}}, \code{\link{getAllCookies}}
}

