% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R, R/documentHandlingDoc.R
\name{getPageSource}
\alias{getPageSource}
\title{Get source of last page.}
\usage{
getPageSource(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}
}
\value{
read_html(res$value): An xml document is returned.
  The content from the remote webdriver is parsed (see \code{\link{read_html}}) and returned
  as xml.
}
\description{
\code{getPageSource} Get the source of the last loaded page.
}
\examples{

}
\seealso{
Other documentHandling functions: \code{\link{executeAsyncScript}},
  \code{\link{executeScript}}
}

