% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentHandling.R, R/documentHandlingDoc.R
\name{getPageSource}
\alias{getPageSource}
\title{getPageSource}
\usage{
getPageSource(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
getPageSource
}
\seealso{
Other documentHandling functions: \code{\link{executeAsyncScript}},
  \code{\link{executeScript}}
}

