% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{getTitle}
\alias{getTitle}
\title{getTitle}
\usage{
getTitle(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
getTitle
}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getCurrentUrl}},
  \code{\link{go}}, \code{\link{refresh}}
}

