% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R, R/navigationDoc.R
\name{getTitle}
\alias{getTitle}
\title{Get the current page title.}
\usage{
getTitle(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
The title of the current page is returned as a character string.
}
\description{
\code{getTitle} Get the current page title.
}
\examples{

}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getCurrentUrl}},
  \code{\link{go}}, \code{\link{refresh}}
}

