% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R
\name{getWindowHandle}
\alias{getWindowHandle}
\title{getWindowHandle}
\usage{
getWindowHandle(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
getWindowHandle
}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

