% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{getWindowHandleOld}
\alias{getWindowHandleOld}
\title{getWindowHandleOld}
\usage{
getWindowHandleOld(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
getWindowHandleOld
}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandlesOld}},
  \code{\link{sendAlertTextOld}}
}

