% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R, R/commandContextsDoc.R
\name{getWindowHandles}
\alias{getWindowHandles}
\title{Get all window handles.}
\usage{
getWindowHandles(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
Returns a list of windows handles. Each element of the list is a string. The order
   window handles are returned is arbitrary.
}
\description{
\code{getWindowHandles} Retrieve the list of all window handles available to the session.
}
\examples{

}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

