% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{getWindowHandlesOld}
\alias{getWindowHandlesOld}
\title{Get all window handles.}
\usage{
getWindowHandlesOld(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
Returns a list of windows handles. Each element of the list is
   a string. The order window handles are returned is arbitrary.
}
\description{
\code{getWindowHandlesOld} Uses the old JSONwireprotocol end points.
   Retrieve the list of all window handles available to the session.
}
\examples{
\dontrun{
# functions in this group are using the old JSONwireprotocol end points
}

}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowPositionOld}},
  \code{\link{getWindowSizeOld}},
  \code{\link{maximizeWindowOld}},
  \code{\link{sendAlertTextOld}},
  \code{\link{setWindowPositionOld}},
  \code{\link{setWindowSizeOld}}
}

