% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R, R/oldMethodsDoc.R
\name{getWindowHandlesOld}
\alias{getWindowHandlesOld}
\title{getWindowHandlesOld}
\usage{
getWindowHandlesOld(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
ret2 : res$value NEED TO FILL IN specifics for each function
}
\description{
\code{getWindowHandlesOld}
}
\examples{

}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowPositionOld}},
  \code{\link{getWindowSizeOld}},
  \code{\link{sendAlertTextOld}},
  \code{\link{setWindowSizeOld}}
}

