% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R, R/commandContextsDoc.R
\name{getWindowPosition}
\alias{getWindowPosition}
\title{Get current window position}
\usage{
getWindowPosition(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
Returns a list which contains the x coordinate to position the window at, relative to the upper left corner of the screen and
 the Y coordinate to position the window at, relative to the upper left corner of the screen
}
\description{
\code{getWindowPosition} The Get Window Position command returns the position on the screen of the operating system window corresponding to the current top-level browsing context.
}
\examples{

}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

