% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{go}
\alias{go}
\title{go}
\usage{
go(remDr, url, ...)
}
\arguments{
\item{remDr}{}
}
\description{
go
}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getCurrentUrl}},
  \code{\link{getTitle}}, \code{\link{refresh}}
}

