% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementState.R
\name{isElementEnabled}
\alias{isElementEnabled}
\title{isElementEnabled}
\usage{
isElementEnabled(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
isElementEnabled
}
\seealso{
Other elementState functions: \code{\link{getElementAttribute}},
  \code{\link{getElementCssValue}},
  \code{\link{getElementProperty}},
  \code{\link{getElementRect}},
  \code{\link{getElementTagName}},
  \code{\link{getElementText}},
  \code{\link{isElementSelected}}
}

