% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{newSession}
\alias{newSession}
\title{newSession}
\usage{
newSession(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
newSession
}
\seealso{
Other sessions functions: \code{\link{deleteSession}},
  \code{\link{setTimeout}}
}

