% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R, R/interactionsDoc.R
\name{performActions}
\alias{performActions}
\title{Not currently implemented}
\usage{
performActions(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently passes the \code{\link{retry}} argument.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{performActions} The Perform Actions command allows you to create sequential interactions that can be sent over from the local end to the remote end. This type of interactions allow emulations like drag and drop.
}
\examples{
\dontrun{
# functions not currently implemented
}

}
\seealso{
Other interactions functions: \code{\link{releasingActions}}
}

