% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R, R/interactionsDoc.R
\name{performActions}
\alias{performActions}
\title{performActions}
\usage{
performActions(remDr, ...)
}
\arguments{
\item{remDr}{remote driver doc}
}
\description{
performActions
}
\seealso{
Other interactions functions: \code{\link{releasingActions}}
}

