% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{refresh}
\alias{refresh}
\title{refresh}
\usage{
refresh(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
refresh
}
\seealso{
Other navigation functions: \code{\link{back}},
  \code{\link{forward}}, \code{\link{getCurrentUrl}},
  \code{\link{getTitle}}, \code{\link{go}}
}

