% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R, R/interactionsDoc.R
\name{releasingActions}
\alias{releasingActions}
\title{Not currently implemented}
\usage{
releasingActions(remDr, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see
\code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently passes the
\code{\link{retry}} argument.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly
   returned. A remote driver object see \code{\link{remoteDr}}. This
   allows for chaining from this function to other functions that take
   such an object as an argument. See examples for further details.
}
\description{
\code{releasingActions} The Release Actions command is used to cancel
   all current action chains. This is the equivalent of releasing all
   modifiers from input sources.
}
\examples{
\dontrun{
# functions not currently implemented
}

}
\seealso{
Other interactions functions: \code{\link{performActions}}
}

