% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{remoteDr}
\alias{remoteDr}
\title{remoteDr}
\usage{
remoteDr(remoteServerAddr = "http://localhost", port = 4444L,
  browserName = "firefox", version = "", platform = "ANY",
  javascript = TRUE, nativeEvents = TRUE, extraCapabilities = list(),
  path = "wd/hub", newSession = TRUE)
}
\description{
remoteDr
}
\examples{

}

