% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldMethods.R
\name{sendAlertTextOld}
\alias{sendAlertTextOld}
\title{sendAlertTextOld}
\usage{
sendAlertTextOld(remDr, ...)
}
\arguments{
\item{remDr}{}
}
\description{
sendAlertTextOld
}
\seealso{
Other oldMethods functions: \code{\link{acceptAlertOld}},
  \code{\link{dismissAlertOld}},
  \code{\link{executeAsyncScriptOld}},
  \code{\link{executeScriptOld}},
  \code{\link{getAlertTextOld}},
  \code{\link{getWindowHandleOld}},
  \code{\link{getWindowHandlesOld}}
}

