% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R, R/sessionsDoc.R
\name{setTimeout}
\alias{setTimeout}
\title{setTimeout}
\usage{
setTimeout(remDr, type = "page load", milliseconds = 10000, ...)
}
\arguments{
\item{remDr}{remote driver doc}

\item{milliseconds}{}
}
\description{
setTimeout
}
\seealso{
Other sessions functions: \code{\link{deleteSession}},
  \code{\link{newSession}}
}

