% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R, R/commandContextsDoc.R
\name{switchToFrame}
\alias{switchToFrame}
\title{switchToFrame}
\usage{
switchToFrame(remDr, Id = NULL, ...)
}
\arguments{
\item{remDr}{}
}
\description{
switchToFrame
}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}},
  \code{\link{switchToParentFrame}},
  \code{\link{switchToWindow}}
}

