% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandContexts.R, R/commandContextsDoc.R
\name{switchToWindow}
\alias{switchToWindow}
\title{switchToWindow}
\usage{
switchToWindow(remDr, name, ...)
}
\arguments{
\item{remDr}{An object of class "rDriver". A remote driver object see \code{\link{remoteDr}}.}

\item{...}{Additonal function arguments - Currently unused.}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{switchToWindow}
}
\examples{

}
\seealso{
Other commandContexts functions: \code{\link{closeWindow}},
  \code{\link{fullscreenWindow}},
  \code{\link{getWindowHandles}},
  \code{\link{getWindowHandle}},
  \code{\link{getWindowPosition}},
  \code{\link{getWindowSize}},
  \code{\link{maximizeWindow}},
  \code{\link{setWindowPosition}},
  \code{\link{setWindowSize}}, \code{\link{switchToFrame}},
  \code{\link{switchToParentFrame}}
}

