% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenCapture.R, R/screenCaptureDoc.R
\name{takeElementScreenshot}
\alias{takeElementScreenshot}
\title{takeElementScreenshot}
\usage{
takeElementScreenshot(webElem, ...)
}
\arguments{
\item{webElem}{web Element doc}
}
\value{
invisible(remDr): An object of class "rDriver" is invisibly returned.
   A remote driver object see \code{\link{remoteDr}}. This allows for chaining from this function to
   other functions that take such an object as an argument. See examples for further details.
}
\description{
\code{takeElementScreenshot}
}
\examples{

}
\seealso{
Other screenCapture functions: \code{\link{takeScreenshot}}
}

