% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenCapture.R, R/screenCaptureDoc.R
\name{takeElementScreenshot}
\alias{takeElementScreenshot}
\title{takeElementScreenshot}
\usage{
takeElementScreenshot(webElem, ...)
}
\arguments{
\item{webElem}{}
}
\description{
takeElementScreenshot
}
\seealso{
Other screenCapture functions: \code{\link{takeScreenshot}}
}

